+++ CHANGELOG.rst
@@ -1,3 +1,12 @@
+0.1.1 -
+---------------
+
+- ``pylint-command`` option added [miohtama]
+
+- Ghetto-CI continuous integration service [miohtama]
+
+- Configuration file reader refactored [miohtama]
+
 0.1 - 0.1.1
 --------------